Add death component to the enemy, add a new death animation, refactor death component
