+++ Assets/Codebase/Logic/IDeathable.cs
@@ -1,0 +1,6 @@
+﻿namespace Codebase.Logic
+{
+  public interface IDeathable
+  {
+  }
+}
