+++ Assets/Codebase/Logic/IDeathable.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d167afdd162c4c7caf1454f77a03c415
+timeCreated: 1655806421
