+++ Assets/Resources/Enemy/Animations/Clips/Death.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 764dde4a73f04604788f8d7786a4e139
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
