+++ Assets/Resources/Enemy/Enemy.prefab
@@ -31,7 +31,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 8700730481765138173}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!135 &8302769281571968927
 SphereCollider:
@@ -92,7 +92,7 @@ RectTransform:
   m_Children:
   - {fileID: 5315789904932125149}
   m_Father: {fileID: 8700730481765138173}
-  m_RootOrder: 7
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -228,6 +228,52 @@ MonoBehaviour:
     m_HintWeight: 1
     m_MaintainTargetPositionOffset: 0
     m_MaintainTargetRotationOffset: 0
+--- !u!1 &3297176853757351082
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1798770776764485262}
+  - component: {fileID: 2297639721540530677}
+  m_Layer: 0
+  m_Name: Rig_Layer_Body_Aim
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1798770776764485262
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3297176853757351082}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0.05295604, y: 1.3555636, z: 0.2081909}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 8700730481765138173}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2297639721540530677
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3297176853757351082}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 70b342d8ce5c2fd48b8fa3147d48d1d1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Weight: 1
+  m_Effectors: []
 --- !u!1 &3342946875649257124
 GameObject:
   m_ObjectHideFlags: 0
@@ -260,7 +306,7 @@ Transform:
   - {fileID: 6412383201359763922}
   - {fileID: 1608707400609006044}
   m_Father: {fileID: 8700730481765138173}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &5540786727382938970
 MonoBehaviour:
@@ -307,7 +353,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 8700730481765138173}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!135 &6607948970368131283
 SphereCollider:
@@ -396,7 +442,7 @@ Transform:
   m_Children:
   - {fileID: 553031648071754243}
   m_Father: {fileID: 8700730481765138173}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1298340108861911923
 MonoBehaviour:
@@ -1127,6 +1173,7 @@ GameObject:
   - component: {fileID: 464021337797881366}
   - component: {fileID: 9208599148623323857}
   - component: {fileID: 4811497424584519527}
+  - component: {fileID: 152245114403450916}
   - component: {fileID: 4756266155227607117}
   m_Layer: 9
   m_Name: Enemy
@@ -1150,6 +1197,7 @@ Transform:
   - {fileID: 8700558591786512089}
   - {fileID: 8700801475829870115}
   - {fileID: 4621220251296508410}
+  - {fileID: 1798770776764485262}
   - {fileID: 6378417486763452170}
   - {fileID: 7552361788298867653}
   - {fileID: 1459466981077350339}
@@ -1243,6 +1291,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_RigLayers:
+  - m_Rig: {fileID: 2297639721540530677}
+    m_Active: 1
   - m_Rig: {fileID: 1298340108861911923}
     m_Active: 1
   - m_Rig: {fileID: 5540786727382938970}
@@ -1317,6 +1367,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _max: 20
+--- !u!114 &152245114403450916
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8694855099473585587}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: edfd02ab45834f88a44c5c2e728e7880, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &4756266155227607117
 MonoBehaviour:
   m_ObjectHideFlags: 0
